CSHARP-6017: Rename LeftJoin tests to QueryableLeftJoinTests and merge serializer finder cases into EnumerableTests
